--- FormsMigrations/UserInterface/ZZZ/WebView/WebView/App.xaml.cs
@@ -1,17 +1,0 @@
-﻿using Microsoft.Maui;
-using Microsoft.Maui.Controls;
-using Microsoft.Maui.Controls.PlatformConfiguration.WindowsSpecific;
-using Application = Microsoft.Maui.Controls.Application;
-
-namespace WebView
-{
-	public partial class App : Application
-	{
-		public App()
-		{
-			InitializeComponent();
-
-			MainPage = new MainPage();
-		}
-	}
-}
