--- FormsMigrations/UserInterface/ZZZ/WebView/WebView/MainPage.xaml.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using Microsoft.Maui.Controls;
-
-namespace WebView
-{
-	public partial class MainPage : ContentPage
-	{
-		public MainPage()
-		{
-			InitializeComponent();
-		}
-
-		int count = 0;
-		private void OnCounterClicked(object sender, EventArgs e)
-		{
-			count++;
-			CounterLabel.Text = $"Current count: {count}";
-		}
-	}
-}
